Enhancement: Windfury Totem-swapping #5
The tick mechanism of Windfury got changed so that it now displays the
uptime of the Windfury Buff instead of the tick between the Totem
applying the Buff.

In addition to that it is now possible to swap the Windfury totem to
other Air Totems and it will still display the current uptime of
Windfury.